+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Create a random user ID for the application
 * Additionally, store a task ID and a task user ID to link to respective surveys
 * If present: delete task ID and task user ID when user hits add-to-cart button
+* Create a `TaskEvent` object that can be send to analytics endpoint
 
 # v0.0.3 | 09.04.2023
 * Attach bages to `AmazonSearchItem` object